Lasd

+++ models/flower.go
@@ -1,10 +1,11 @@
 package models
 
+//struktur tabel `flower`
 type Flower struct {
 	Id           int
 	Name         string
 	Category     string
-	Price        float32
+	Price        float64
 	Photo        string
 	Descriptions string
 }
